@@ -1,16 +1,15 @@
 name: graphs
-version: 0.2.1-dev
+version: 0.3.0-nullsafety.0
 description: Graph algorithms that operate on graphs in any representation
 homepage: https://github.com/dart-lang/graphs
 
 environment:
-  sdk: '>=2.3.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dev_dependencies:
-  pedantic: ^1.3.0
-  test: ^1.5.1
+  pedantic: ^1.10.0-nullsafety.3
+  test: ^1.16.0-nullsafety.13
 
   # For examples
-  analyzer: '>=0.35.0 <0.42.0'
-  path: ^1.1.0
-  pool: ^1.3.0
+  path: ^1.8.0-nullsafety.3
+  pool: ^1.5.0-nullsafety.3
