@@ -34,7 +34,7 @@ final _empty = Future<Null>.value(null);
 /// have not completed. If the [edges] callback needs to be limited or throttled
 /// that must be done by wrapping it before calling [crawlAsync].
 Stream<V> crawlAsync<K, V>(Iterable<K> roots, FutureOr<V> Function(K) readNode,
-    FutureOr<Iterable<K>> Function(K, V) edges) {
+    FutureOr<Iterable<K>?> Function(K, V) edges) {
   final crawl = _CrawlAsync(roots, readNode, edges)..run();
   return crawl.result.stream;
 }
@@ -43,7 +43,7 @@ class _CrawlAsync<K, V> {
   final result = StreamController<V>();
 
   final FutureOr<V> Function(K) readNode;
-  final FutureOr<Iterable<K>> Function(K, V) edges;
+  final FutureOr<Iterable<K>?> Function(K, V) edges;
   final Iterable<K> roots;
 
   final _seen = HashSet<K>();
